Move address-formatter back to its original package, add license and readme
