+++ libs/address-formatter/src/main/java/org/woheller69/RegexPatternCache.java
@@ -1,4 +1,4 @@
-package de.mm20.launcher2.addressformatter;
+package org.woheller69;
 
 import java.util.Map;
 import java.util.HashMap;
